Updated the logic of the form to allow reporting without UPRN.
User is forced to fill in the location description in this scenario and select the propery type themselves.

You are still presented with the list of addresses based on the postcode of the location you selected and if you select an address it defaults back to the standard with address flow.

This is a WIP commit as we may need to redesign.
Main issues are:
 - Should we be using the Address.razor page or spliting to seperate locations?
 - Do we need to include a map image above the location box

Tested flow from beginning to end without saving in all contexts and believe you can get to summary page.